Employee_Agent_Foundry/requirements.txt
python-dotenv>=1.0.1
azure-identity>=1.15.0
aiohttp>=3.9.0
azure-core>=1.29.0

azure-ai-projects
azure-ai-agents

agent-framework
agent-framework-azure-ai